Convert Channels into an array
Profiling showed that over half (!) of the time is spent doing
SortedList key lookups on SortedChannels.  In contrast, array indexing
can be optimized into pointer arithmetic, so it is much more efficient.

This results in a 2x performance improvement for both loading and
saving.  The remaining hotspots are in color-conversion and MemoryBlock
accesses.