Process layers/channels in parallel
The last change caused a doubling of PSD saving time, as RLE compression
is a processor-intensive operation.  To counteract this performance
drop, I have used the Paint.NET private threadpool to parallelize most
of the loading and saving operations at the granularity of layers or
channels.  On a system with lots of cores, the plugin now performs quite
well indeed.

There is still some sequentiality remaining (notably in masks).  Also,
single-layer images will only benefit from the parallelization that was
done at a channel granularity.  However, it's mostly the images with
high layer counts that tend to slow down anyway.