Remove the stand-alone PSD code and test app
The stand-alone PSD library was largely *duplicating* the equivalent
code in the PSD plugin code.  However, I have not been maintaining it
consistently.  Since it was in the source repository, though, some
developers have been attempting to use this code and have been running
into problems.

In fact, I also got confused and added the bitmap helper functions to
ImageDecoderBitmaps.  This has been moved to ImageDecoderPdn and to a
separate Util class.

Although it would be useful to have a DLL for PSD files that can be used
outside of Paint.NET, duplicate code is not the right way to accomplish
this.