Check for enough memory to load into Paint.NET
This is the number one generator of user inquiries.

Photoshop files with lots of small adjustment layers will expand
dramatically in Paint.NET, where layers have the same dimensions as the
canvas.  When the system starts paging to/from swap after physical
memory exhaustion, it looks like a hang to users.