Improved algorithm for finding layer bounds
Instead of scanning through the entire image for nontransparent pixels,
the algorithm takes advantage of previously-found pixels to limit its
search for later rows.  This substantially improves saving speed for
images that take up all or most of a layer, e.g., photos.

This change also fixes the crash that previously occurred when saving
empty layers.