Refactor out ChannelsList, clean up constructors
The Layer had been maintaining the channels as two lists, a SortedList
and an array, which might get out of sync.  The ChannelsList 't have to
worry about maintaining two lists.

Constructors no longer add themselves to the parent.  This is consistent
with usual practice and eliminates the possiblity of accidental
duplication.