Check virtual memory address space before loading
The plugin currently refuses to load any file that exceeds the available
physical memory.  However, 32-bit systems may have more physical memory
than can be accessed by any user-mode process, as some of the 32-bit
virtual address space is reserved for the kernel.

This does not hang, but gives the "unspecified error" dialog.  This is
because the memory allocations fail for code that is running inside the
Paint.NET PrivateThreadPool, which wraps exceptions.  Paint.NET used to
unwrap the exception in older versions, but does not do so in the
current version (3.5.10).  Thus, it is not recognized as an
OutOfMemoryException, and the catch-all error dialog is displayed.

We can avoid this problem by checking virtual memory outside of the
threadpool, and throwing an OutOfMemoryException ourselves.  Note that
virtual memory maxes out at 2 GB, even if the user has increased the
user-mode virtual address space, since PaintDotNet.exe does not have
/LARGEADDRESSAWARE set.