Rewrite of the layer mask code
The code now handles many more features of masks.  It calculates the
correct position when clipping to the mask rectangle.  Portions of the
layer that fall outside of the layer mask are filled using the
background mask color.

Both mask channels (-2 and -3) are handled and applied during image
decoding.  This includes the vector mask, if a rasterized version of the
mask is present in channel -2.

Unfortunately, most complex PSD files still cannot be masked properly,
as a rasterized version of the vector mask is usually (but not always)
omitted by Photoshop.