Work item 7966: Load multichannel files
Multichannel files only have one layer.  This allows each channel to be
mapped to a layer in a Paint.NET document, using the channel name as the
layer name.

The primary purpose of multichannel files is to represent the plates
used in offset printing.  Thus, it is best to keep the channels data
separate, so that they can be accessed individually.