Speed up decoding using delegates and inlining
Previously, SetPDNColorRow called SetPDNColor in a loop, and SetPDNColor
then used a switch-case based on the color mode of the image.  Thus, we
were incurring a function call overhead and a switch-case on every
pixel.

Now, the switch-case is located outside of the loop, and is used to
select a DecodeDelegate that matches the color mode and the bitdepth.
Each DecodeDelegate does its own inner loop over the pixels in a row,
and the color-conversion code is in the body of the loop.  This
eliminates both the switch-case and the function call overhead.

Image decoding gets a speedup of 1.1x to 1.5x on 8-bit images.  RGB sees
the greatest speedup factor.  Lab images see the least speedup factor,
with some performance gains coming from eliminating unnecessary floating
point operations.  32-bit images already had a tight inner loop, but a
15% speedup has been achieved by shaving RGBByteFromHDRFloat to fit
within the 32-byte IL limit for inlining.

The use of delegates improves the organization of the code, and keeps
the switch-case out of the outer loop in DecodeImage.  The delegate
signature has been kept manageable by collecting parameters together
into the DecodeContext.