Issue 7978: RleWriter inconsistency
The RleWriter previously encoded all two-byte runs as RLE.  This was
inefficient, as the run may end after two bytes and waste a byte to
return to a raw packet context.

Now, the RleWriter behaves the same as the PackBits RLE implementation
in Photoshop.  Raw packets are not interrupted until a three-byte run is
encountered, but two-byte runs are otherwise still stored as RLE.

Average compressed data size has shrunk from 101.3% of Photoshop to
97.7%, with all test cases <= Photoshop.  This is because Photoshop does
not always max out all 128 bytes of a repeat packet, so it sometimes has
to use more packets to represent a long run.  However, files still
increase in size when going through Paint.NET, as Paint.NET renders out
a less-compressible composite image.