Simplify accessors for channel image data
The raw and decompressed image data are now exposed directly to
accessors.

The flags were an attempt to keep the two arrays in sync.  However, they
made the logic harder to follow, and did not protect against the common
case of modifying the image data in-place.  It is better to remove the
flags and have the caller keep the image data in sync.