Issue 7985: Incorrect padding on layer info
A regression was introduced on 2013-02-08 in attempting to replicate
Photoshop behavior when padding layer info.

Photoshop sometimes writes a different length than the actual length.
However, files created by previous versions of the PSDPlugin did not do
this, so they fail to load into the newest release of the PSDPlugin.

The undocumented padding behavior is now restricted to the global layer
info.  Photoshop does not have a length mismatch for layer info at the
end of each layer record.