Improve debug messages
By printing a debug message at the end of a section as well as the
beginning, we get a better sense of the structure of the PSD file.  The
use of a consistent CSV format makes it possible to parse the messages,
and the hex position is convenient when examining files in a hex editor.