Use TPL and PLINQ for parallel processing
In Paint.NET 3.5.x, we used the PrivateThreadPool for parallel
processing, because the Parallel Extensions for .NET 3.5 required a
separate install.

Now that Paint.NET is on .NET 4, we can count on having the TPL and
PLINQ in the base install.  By using standard .NET features for parallel
code, we eliminate the dependency on the PrivateThreadPool.  The code is
also cleaner because the callbacks and context classes are not needed.