Simplify layer and mask decoding
Rectangle.Intersect is used to calculate the decode rectangles for
layers and masks.  This simplifies the code and makes the arithmetic
easier to follow.

We now loop only on y when decoding the layer, calculating relative
coordinates within the loop rather than incrementing multiple loop
variables.  For clarity, x and y are reserved for absolute coordinates,
while row and column are used for relative coordinates.

While decoding a mask, the alpha buffer is allocated per-layer and
reused for each row, and the color mode selection is done outside of the
inner loop.  In addition, we no longer offset the relative mask position
by the layer position, since Photoshop does not do this.