Early exit if file doesn't fit in memory
The memory required for the Paint.NET Document is now checked as soon as
the number of layers is known.  An OutOfMemoryException can then be
thrown early, skipping the rest of the file.

This is important for large uncompressed PSB files.  Loading in a multi-
gigabyte file could take a long time and force most of memory to be
paged out to disk.  It is much less disruptive to exit early.