Save files that are not 8-bits per channel
Layers can now be saved to the Additional Info section rather than the
Layers Info section.  They are still stored in the PsdFile class, but
the presence of an InfoLayers object triggers the different behavior
during save.

Photoshop saves Additional Info layers for images with 16 or 32 bits
per channel.  These higher-bitdepth files can now be loaded and saved to
be bit-for-bit identical.  Saving of black-and-white files is also
enabled in this revision.