Flush PsdBinaryWriter before returning BaseStream
An accurate Stream.Position is required for the PsdBlockLengthWriter and
the debug messages to work properly.  Therefore, the PsdBinaryWriter
must be flushed before returning the BaseStream.

This removes the need for the AutoFlush property.  The caller should be
responsible for properly disposing of the PsdBinaryWriter, just like
with the .NET BinaryWriter.  The AutoFlush property was not being
checked in all Write overloads anyway.