Simplify pointer operations in Unpredict methods
Using UInt16* and Int32* pointers removes the need to cast from byte*.
This eliminates the issue of whether to add a byte offset or a word
offset to get a pointer to the new row.  There is also semantic value in
knowing that the Unpredict method operates on words instead of bytes.

The pointer parameters are used directly, reducing the number of
temporary variables.  Switching to bit-shifts for 32-bit Unpredict
eliminates the need to advance pointers while unpacking each pixel, and
it also makes the code endian-neutral.