Skip global layer mask if section ends early
When there is no global layer mask, Photoshop writes out a length of
zero.  However, some programs end the section early when there are also
no LayerInfo records.  We must check the remaining length to avoid
reading past the end of the Layer and Mask Information section.

This is a longstanding bug that had been masked by resetting the stream
position at the end of the section.  The last revision unmasked the bug
by removing the workaround and checking for overruns.